fix(nlsq): prevent int32 overflow in flat index computation
Replace int32 with int64 for xdata arrays and searchsorted results
used in flat index computation (phi_idx * n_t1 * n_t2 + ...). For
large datasets (>2.1B points), int32 wraps to negative values,
silently corrupting model evaluation. Fixes 16 sites across 5 files:
wrapper.py (3), parallel_accumulator.py (6), residual.py (4),
residual_jit.py (3), fitting.py (1).  [QA round 4]

+++ homodyne/core/fitting.py
@@ -780,7 +780,7 @@ if JAX_AVAILABLE:
             jnp.array(0.0, dtype=jnp.float64),  # sum_theory
             jnp.array(0.0, dtype=jnp.float64),  # sum_exp
             jnp.array(0.0, dtype=jnp.float64),  # sum_theory_exp
-            jnp.array(0, dtype=jnp.int32),  # n_data (already fixed in R6)
+            jnp.array(0, dtype=jnp.int64),  # n_data (R6: int→jnp, R4-4: int32→int64)
         )
 
         # Process all chunks
